re-structuring meta, base and factory code
