+++ jumpscale/clients/github/__init__.py
@@ -1,4 +1,4 @@
-from jumpscale.clients.base import StoredFactory
+from jumpscale.core.base import StoredFactory
 from .github import Github
