+++ jumpscale/core/base/__init__.py
@@ -1,0 +1,2 @@
+from .meta import Base
+from .factory import StoredFactory
